requirements.txt
Django<4.3
djangorestframework<3.16
psycopg2
drf-spectacular
